[cmake] Introduce MC_RTC_HONOR_INSTALL_PREFIX
This CMake variable modifies the behavior of mc_rtc macros so that the
install happens in the specified CMAKE_INSTALL_PREFIX rather than
mc_rtc's install prefix.

This is especially useful to install mc_rtc component on a system where
on has no write access to mc_rtc install location.